Chat Bot - Code commit
-Code relate to chat bot functionality
-Fix related to discount
-Fix related to sort

+++ static/js/sort.js
@@ -12,8 +12,8 @@ function sort(sortBy){
 	const sortedItems = Array.from(gridItems).sort(function(a,b){
 		
 		if (sortBy == 'plh' || sortBy == 'phl'){
-			const priceA = parseInt(a.querySelector(".labelPrice").getAttribute('data-price'));
-			const priceB = parseInt(b.querySelector(".labelPrice").getAttribute('data-price'));
+			const priceA = parseFloat(a.querySelector(".labelPrice").getAttribute('data-price'));
+			const priceB = parseFloat(b.querySelector(".labelPrice").getAttribute('data-price'));
 			switch (sortBy) {
 				case 'plh':
 					return priceA - priceB
@@ -23,8 +23,8 @@ function sort(sortBy){
 					break;
 			}
 		}else if (sortBy =='nwa') {
-				const dateA = parseInt(a.querySelector(".labelDate").getAttribute('data-date'));
-				const dateB = parseInt(b.querySelector(".labelDate").getAttribute('data-date'));
+				const dateA = parseFloat(a.querySelector(".labelDate").getAttribute('data-date'));
+				const dateB = parseFloat(b.querySelector(".labelDate").getAttribute('data-date'));
 				return dateB - dateA;
 
  		}
